chore: rename package builder to plugin
